# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raumdeuter API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/potential": {
      "post": {
        "tags": [
          "potential"
        ],
        "summary": "Potential",
        "operationId": "potential_potential_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PotentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PotentialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/potential/cohort": {
      "post": {
        "tags": [
          "potential"
        ],
        "summary": "Potential Cohort",
        "description": "Paginated age-cohort slice \u2014 ordered by ``potential_score`` desc, with total count.",
        "operationId": "potential_cohort_potential_cohort_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PotentialCohortRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PotentialCohortResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerFilters": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season",
            "description": "Season start year, e.g. 2025 for 25-26"
          },
          "leagues": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leagues"
          },
          "roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roles"
          },
          "teams": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teams"
          },
          "age_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Min"
          },
          "age_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Max"
          },
          "minutes_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes Min"
          },
          "minutes_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes Max"
          },
          "market_value_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value Max"
          },
          "foot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foot"
          },
          "contract_expires_year_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Expires Year Min"
          },
          "contract_expires_year_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Expires Year Max"
          }
        },
        "type": "object",
        "required": [
          "season"
        ],
        "title": "PlayerFilters"
      },
...
      "PotentialCohortRequest": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/PlayerFilters"
          },
          "age": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 14.0,
            "title": "Age"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          },
          "offset": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "filters",
          "age"
        ],
        "title": "PotentialCohortRequest",
        "description": "Single age-cohort slice for the Potential big-board (paginated)."
      },
      "PotentialCohortResponse": {
        "properties": {
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PotentialPlayer"
            },
            "type": "array",
            "title": "Players"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "age",
          "players",
          "total"
        ],
        "title": "PotentialCohortResponse"
      },
      "PotentialPlayer": {
        "properties": {
          "wyscout_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wyscout Id"
          },
          "player": {
            "type": "string",
            "title": "Player"
          },
          "club": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club"
          },
          "club_logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club Logo"
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "current_pi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Pi"
          },
          "potential_score": {
            "type": "number",
            "title": "Potential Score"
          },
          "player_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Image Url"
          }
        },
        "type": "object",
        "required": [
          "wyscout_id",
          "player",
          "potential_score"
        ],
        "title": "PotentialPlayer"
      },
      "PotentialRequest": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/PlayerFilters"
          }
        },
        "type": "object",
        "required": [
          "filters"
        ],
        "title": "PotentialRequest"
      },
      "PotentialResponse": {
        "properties": {
          "players": {
            "items": {
              "$ref": "#/components/schemas/PotentialPlayer"
            },
            "type": "array",
            "title": "Players"
          },
          "n": {
            "type": "integer",
            "title": "N"
          }
        },
        "type": "object",
        "required": [
          "players",
          "n"
        ],
        "title": "PotentialResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}